vbus: fix various warnings
There are various warnings compiling the vbus code on x86. Fix the easiest
of them.

+++ kernel/vbus/config-ioctl.c
@@ -207,7 +207,7 @@ out:
 	return ret;
 }
 
-static ssize_t
+static long
 vbus_admin_chardev_ioctl(struct file *filp, unsigned int ioctl,
 			 unsigned long arg)
 {
